Stochastic lesson to quarto.yml
Added stochastic lesson to _quarto.yml and rendered book
